+++ SortingData.hpp
@@ -1,14 +1,16 @@
 #ifndef ED_TP2_SORTINGDATA_HPP
 #define ED_TP2_SORTINGDATA_HPP
 
+#include <cstdint>
+
 class SortingData {
 private:
-    int comparisions;
-    int swaps;
+    uint64_t comparisions;
+    uint64_t swaps;
 public:
     SortingData();
-    int getComparisons();
-    int getSwaps();
+    uint64_t getComparisons();
+    uint64_t getSwaps();
     void incrementComparisons();
     void incrementSwaps();
 };
